refactor: split out save and load
